pyproject.toml
[project]
name = "alibaba-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "deepagents>=0.5.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.1.14",
    "langchain-qwq>=0.3.4",
    "markdownify>=1.2.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "sentence-transformers>=5.4.1",
    "opencv-python>=4.10.0",
]
